got bcexmm to compile again. i got alot of work todo with this section of the code but the core is important af

+++ src/util/bcexmm.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
 if [ -f "BCEXMM.MAP" ]; then
-	mv BCEXMM.EXE bcexmm.ex0
-	mv bcexmm.ex0 bcexmm.exe
+	if [ -f "BCEXMM.EXE" ]; then
+		mv BCEXMM.EXE bcexmm.ex0
+		mv bcexmm.ex0 bcexmm.exe
+	fi
 	mv BCEXMM.MAP bcexmm.meh
 fi
